+++ src/v1/BasicTypes/Primitives/index.ts
@@ -32,20 +32,7 @@
 // POSSIBILITY OF SUCH DAMAGE.
 //
 
-export * from "./Any";
-export * from "./Arrays";
-export * from "./Booleanish";
-export * from "./Booleans";
-export * from "./Classes";
-export * from "./Filters";
-export * from "./Functions";
-export * from "./HashMaps";
-export * from "./Integers";
-export * from "./Numbers";
-export * from "./Objects";
-export * from "./Options";
-export * from "./Primitives";
-export * from "./Prototypes";
-export * from "./Regex";
-export * from "./Strings";
-export * from "./Unknowns";
+export * from "./Primitive";
+export * from "./PrimitiveOrFunction";
+export * from "./PrimitiveOrUndefined";
+export * from "./Primitivish";