+++ CHANGELOG.md
@@ -21,6 +21,14 @@ For each release, changes are grouped under these headings:
 
 The following changes have been completed, and will be included in the next tagged release.
 
+### New
+
+* Basic Types
+  - added `Primitive`
+  - added `PrimitiveOrFunction`
+  - added `PrimitiveOrUndefined`
+  - added `Primitivish`
+
 ## v0.2.6
 
 Released Sunday, 4th October 2020.